double to decimal

+++ src/AthenasAcademy.Certificate.Domain/Requests/CertificateRequest.cs
@@ -4,7 +4,7 @@ public record CertificateRequest
 {
     public StudentRequest Student { get; set; }
     public CourseRequest Course { get; set; }
-    public double Utilization { get; set; }
+    public decimal Utilization { get; set; }
     public DateTime ConclusionDate { get; set; }
 
 }
